feat: add ability to temporarily ignore clocking in
This feature exists so that a user can hop into a buffer quickly or a
series of buffers without Emacs pestering the user too much for them
to clock-in.

When a user sends `quit' error when trying to select a task this will
cause the current buffer to be put on an ignore list. This ignore list
prevents org-auto-clock from trying to clock-in to a task for the user
from that buffer. The list only gets cleared out after the first
successful clocking in of a task.